binary functions working

+++ include-templates/fun_binary_functions.h
@@ -16,7 +16,7 @@
 
 // TODO: realtime check that deep dimensions of E1 and E2 are the same
 
-template <class A, class B, class E1, class E2, class NT1, class NT2, int depth, int rank>
+template <class A, class B, class E1, class E2, class NT1, class NT2, size_t depth, size_t rank>
 auto ##FUNCTION##(const ExpressionR<A, E1, NT1, depth, rank>& x1, const ExpressionR<B, E2, NT2, depth, rank>& x2) {
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef typename NumberTrait<E1, NT3>::ReplacedNumberType E3;   // see TODO note above
@@ -31,7 +31,7 @@ auto ##FUNCTION##(const ExpressionR<A, E1, NT1, depth, rank>& x1, const Expressi
 
 // (2A) MultiArray<Element(NT1)> ##OP## NT2
 
-template <class A, class NT2, class Element, class NT1, int depth, int rank, typename = std::enable_if_t<NumberTrait<NT2>::value>>
+template <class A, class NT2, class Element, class NT1, size_t depth, size_t rank, typename = std::enable_if_t<NumberTrait<NT2>::value>>
 auto ##FUNCTION##(const ExpressionR<A, Element, NT1, depth, rank>& x1, const NT2& x2) {
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef typename NumberTrait<Element, NT3>::ReplacedNumberType E3;
@@ -45,7 +45,7 @@ auto ##FUNCTION##(const ExpressionR<A, Element, NT1, depth, rank>& x1, const NT2
 
 // (2B) NT1 ##OP## MultiArray<Element(NT2)>
 
-template <class NT1, class B, class Element, class NT2, int depth, int rank, typename = std::enable_if_t<NumberTrait<NT1>::value>>
+template <class NT1, class B, class Element, class NT2, size_t depth, size_t rank, typename = std::enable_if_t<NumberTrait<NT1>::value>>
 auto ##FUNCTION##(const NT1& x1, const ExpressionR<B, Element, NT2, depth, rank>& x2) {
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef typename NumberTrait<Element, NT3>::ReplacedNumberType E3;
@@ -62,13 +62,13 @@ auto ##FUNCTION##(const NT1& x1, const ExpressionR<B, Element, NT2, depth, rank>
 // TODO: run-time check (deep dimensions of E1 == deepdimensions of x2)
 
 
-template <class A, class B, class E1, class E2, class NT1, class NT2, int D1, int D2, int R1, int R2,
+template <class A, class B, class E1, class E2, class NT1, class NT2, size_t D1, size_t D2, size_t R1, size_t R2,
   EnableIf<(D1==D2+1)&&(IsMultiArrayOrExpression<E1>::value)&&(E1::rank_value==R2)> = 0 >
   auto ##FUNCTION##(const ExpressionR<A, E1, NT1, D1, R1>& x1, const ExpressionR<B, E2, NT2, D2, R2>& x2) {
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef E1 Element;   // see TODO note above
-  constexpr int R3 = R1;
-  constexpr int D3 = D1;
+  constexpr size_t R3 = R1;
+  constexpr size_t D3 = D1;
   typedef typename NumberTrait<Element, NT3>::ReplacedNumberType E3;
   //    MOUT << "A" <<std::endl;
   return  TER_Binary<ExpressionR<A, E1, NT1, D1, R1>,
@@ -82,13 +82,13 @@ template <class A, class B, class E1, class E2, class NT1, class NT2, int D1, in
 
 // TODO: run-time check (deep dimensions of x1 == deepdimensions of E2)
 
-template <class A, class B, class E1, class E2, class NT1, class NT2, int D1, int D2, int R1, int R2,
+template <class A, class B, class E1, class E2, class NT1, class NT2, size_t D1, size_t D2, size_t R1, size_t R2,
   EnableIf<(D1+1==D2)&&(IsMultiArrayOrExpression<E2>::value)&&(E2::rank_value==R1)> = 0 >
   auto ##FUNCTION##(const ExpressionR<A, E1, NT1, D1, R1>& x1, const ExpressionR<B, E2, NT2, D2, R2>& x2) {
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef E2 Element;   // see TODO note above
-  constexpr int R3 = R2;
-  constexpr int D3 = D2;
+  constexpr size_t R3 = R2;
+  constexpr size_t D3 = D2;
   typedef typename NumberTrait<Element, NT3>::ReplacedNumberType E3;
   //    MOUT << "B" <<std::endl;
   return  TER_Binary<ExpressionR<A, E1, NT1, D1, R1>,
@@ -106,7 +106,7 @@ template <class A, class B, class E1, class E2, class NT1, class NT2, int D1, in
 // TODO: if top-level: run-timecheck dimesions of x1  equal dimensions of x2
 // TODO: if element-wise: run-timecheck dimesions of E1  equal dimensions of x2
 
-template <class A, class B, class E1, class NT1, class NT2, int D1, int D2, int rank,
+template <class A, class B, class E1, class NT1, class NT2, size_t D1, size_t D2, size_t rank,
   EnableIf<(D1>=2)&&(D2==1)&&(IsMultiArrayOrExpression<E1>::value)> = 0  >
   auto ##FUNCTION##(const ExpressionR<A, E1, NT1, D1, rank>& x1, const ExpressionR<B, NT2, NT2, D2, rank>& x2) {
 
@@ -117,7 +117,7 @@ template <class A, class B, class E1, class NT1, class NT2, int D1, int D2, int 
 
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef typename NumberTrait<E1, NT3>::ReplacedNumberType E3;   // see TODO note above
-  constexpr int D3 = D1;
+  constexpr size_t D3 = D1;
   //    MOUT << "C" <<std::endl;
   return  TER_Binary<ExpressionR<A, E1, NT1, D1, rank>,
     ExpressionR<B, NT2, NT2, D2, rank>,
@@ -133,7 +133,7 @@ template <class A, class B, class E1, class NT1, class NT2, int D1, int D2, int 
 // TODO: if top-level: run-timecheck dimesions of x1  equal dimensions of x2
 // TODO: if element-wise: run-timecheck dimesions of x1  equal dimensions of E2
 
-template <class A, class B, class E2, class NT1, class NT2, int D1, int D2, int rank,
+template <class A, class B, class E2, class NT1, class NT2, size_t D1, size_t D2, size_t rank,
   EnableIf<(D1==1)&&(D2>=2)&&(IsMultiArrayOrExpression<E2>::value)> = 0  >
   auto ##FUNCTION##(const ExpressionR<A, NT1, NT1, D1, rank>& x1, const ExpressionR<B, E2, NT2, D2, rank>& x2) {
   // NOT SURE WHICH OF THE THREE IS BEST
@@ -142,7 +142,7 @@ template <class A, class B, class E2, class NT1, class NT2, int D1, int D2, int 
   typedef ExpressionR<A, NT1, NT1, D1, rank> E1;
   typedef typename ##TYPECLASS##<NT1, NT2>::Type NT3;
   typedef typename NumberTrait<E2, NT3>::ReplacedNumberType E3;   // see TODO note above
-  constexpr int D3 = D2;
+  constexpr size_t D3 = D2;
   //    MOUT << "Number" <<std::endl;
   return  TER_Binary<ExpressionR<A, NT1, NT1, D1, rank>,
     ExpressionR<B, E2, NT2, D2, rank>,
